+++ src/BasicInputs/StringInput.tsx
@@ -10,6 +10,14 @@ import { Textarea } from '@/lib/components/ui/textarea';
 export const StringInput = forwardRef<
   ElementRef<typeof Textarea>,
   FormControlChildProps & ComponentPropsWithoutRef<typeof Textarea>
->((props, ref) => {
-  return <Textarea rows={1} autoComplete="off" ref={ref} {...props} />;
+>(({ value, ...props }, ref) => {
+  return (
+    <Textarea
+      value={value ?? ''}
+      rows={1}
+      autoComplete="off"
+      ref={ref}
+      {...props}
+    />
+  );
 });