+++ src/Form/FormControl.tsx
@@ -1,17 +1,16 @@
-import { Slot } from '@radix-ui/react-slot';
 import type { HTMLProps, PropsWithoutRef, ReactNode } from 'react';
+
 import {
   type ControllerRenderProps,
   type UseControllerProps,
   useController,
 } from 'react-hook-form';
 
-import type { SelectOption } from '@/types';
+import { Slot } from '@radix-ui/react-slot';
 
 import { useFormStore } from './useFormStore';
 
 export interface FormControlProps extends Omit<UseControllerProps, 'control'> {
-  options?: SelectOption[];
   children: ReactNode;
 }
 
@@ -22,12 +21,10 @@ export interface FormControlChildProps
   'aria-describedby'?: string;
   'aria-invalid'?: boolean;
   'aria-disabled'?: boolean;
-  options?: SelectOption[];
 }
 
 export const FormControl = ({
   name,
-  options,
   children,
   ...controllerProps
 }: FormControlProps) => {
@@ -44,7 +41,6 @@ export const FormControl = ({
   const slotProps = {
     ...field,
     disabled: isSubmitting,
-    options,
   };
   return (
     <Slot