+++ src/ModelTable/stories/ModelTable.stories.tsx
@@ -1,5 +1,6 @@
-import type { Meta, StoryObj } from '@storybook/react';
 import { useState } from 'react';
+
+import type { Meta, StoryObj } from '@storybook/react';
 
 import {
   type ActionParams,
@@ -27,7 +28,7 @@ const meta = {
       {
         id: '1',
         message: 'Make Table Scrollable',
-        user: { id: '00000001', _display_value: 'robxbob' },
+        user: { id: '00000001', displayValue: 'robxbob' },
         created_at: '2024-07-10T01:56:34.926365',
         points: 1,
         done: true,
@@ -35,7 +36,7 @@ const meta = {
       {
         id: '2',
         message: 'Resizable Columns',
-        user: { id: '00000002', _display_value: 'nicklitvin' },
+        user: { id: '00000002', displayValue: 'nicklitvin' },
         created_at: '2024-08-01T01:56:34.926365',
         points: 3,
         done: false,
@@ -43,7 +44,7 @@ const meta = {
       {
         id: '3',
         message: 'Feature: Column DnD',
-        user: { id: '00000001', _display_value: 'robxbob' },
+        user: { id: '00000001', displayValue: 'robxbob' },
         created_at: '2024-07-29T01:56:34.926365',
         points: 4,
         done: true,
@@ -51,7 +52,7 @@ const meta = {
       {
         id: '4',
         message: 'Feature: React Select',
-        user: { id: '00000003', _display_value: 'cmacgray14' },
+        user: { id: '00000003', displayValue: 'cmacgray14' },
         created_at: '2024-08-14T01:56:34.926365',
         points: 2,
         done: false,
@@ -65,13 +66,15 @@ const meta = {
         hidden: true,
       },
       user: {
-        type: FieldType.MODEL_ITEM,
-        valueOptions: [
-          { label: 'robxbob', value: '00000001' },
-          { label: 'nicklitvin', value: '00000002' },
-          { label: 'cmacgray14', value: '00000003' },
-          { label: 'None', value: null },
-        ],
+        type: FieldType.MODEL,
+        inputProps: {
+          options: [
+            { displayValue: 'robxbob', id: '00000001' },
+            { displayValue: 'nicklitvin', id: '00000002' },
+            { displayValue: 'cmacgray14', id: '00000003' },
+            { displayValue: 'None', id: null },
+          ],
+        },
       },
       message: {
         required: true,